# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-Commerce Analytics & Recommendation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/top-selling:
    get:
      operationId: getTopSellingProducts
      tags: [products]
      summary: Get top selling products
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Top selling products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductSalesStat"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: from
          schema:
            type: string
        - in: query
          name: to
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ANALYTICS
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, category, price, stock, rating]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        rating:
          type: number
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, category, price, stock]
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        rating:
          type: number
        description:
          type: string
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        rating:
          type: number
        description:
          type: string
        imageUrl:
          type: string
    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductSalesStat:
      type: object
      required: [id, name, category, totalSold, revenue, rating]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        totalSold:
          type: integer
        revenue:
          type: number
        rating:
          type: number
...
    Order:
      type: object
      required: [id, userId, totalAmount, orderDate, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        orderDate:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
    OrderInput:
      type: object
      required: [userId, items]
      properties:
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    OrderList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...