# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Analytics & Recommendation System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  stock: number;
  rating: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt?: string;
}