# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Analytics & Recommendation System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  category: string;
  price: number;
  stock: number;
  rating?: number;
  description?: string;
  imageUrl?: string;
}