# lib/api-zod/src/generated/types/productSalesStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Analytics & Recommendation System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductSalesStat {
  id: number;
  name: string;
  category: string;
  totalSold: number;
  revenue: number;
  rating: number;
}